Make some tech icons invert colors on dark backgrounds [ci skip]
