+++ src/components/Portfolio/Sections/AboutMe.component.tsx
@@ -18,7 +18,7 @@ import {
 import { ImageQuery } from "@/types/general";
 import { downloadFile } from "@/utils/dom";
 import { GatsbyImage } from "gatsby-plugin-image";
-import { convertHexToHSL, parseHSLString } from "@/utils/colors";
+import { convertHexToHSL, isBgDark, parseHSLString } from "@/utils/colors";
 
 const StyledProfile = styled.div`
 	display: flex;
@@ -153,24 +153,6 @@ const ResumeButton: React.FC<{ resume: string }> = ({ resume }) => (
 		/>
 	</StyledResumeButton>
 );
-
-function isBgDark(id: string, name: string, bgColor: string): boolean {
-	if (id === "1" || name.toLowerCase().includes("night")) {
-		return true;
-	}
-
-	let hsl = parseHSLString(bgColor);
-	if (hsl && hsl.luminance < 50) {
-		return true;
-	}
-
-	try {
-		hsl = convertHexToHSL(bgColor);
-		return hsl.luminance < 50;
-	} catch {
-		return false;
-	}
-}
 
 const AboutMe: React.FC<{
 	ghIcon?: string;
